Remove monkey patch from the main class to fix some documentation issue
